+++ crates/runtime/Cargo.toml
@@ -17,7 +17,6 @@ thiserror.workspace = true
 ureq.workspace = true
 wasmer.workspace = true
 wasmer-types.workspace = true
-borsh = { workspace = true, features = ["derive"] }
 
 [[example]]
 name = "demo"
