+++ crates/sdk/src/env.rs
@@ -1,5 +1,3 @@
-use std::collections::HashMap;
-
 use crate::sys;
 
 #[doc(hidden)]
@@ -156,17 +154,4 @@ pub fn state_write<T: crate::state::AppState>(state: &T) {
         Err(err) => panic_str(&format!("Cannot serialize app state: {:?}", err)),
     };
     storage_write(STATE_KEY, &data);
-}
-
-pub fn fetch(method: &str, url: &str, headers: HashMap<String, String>, body: Vec<u8>) -> String {
-    let headers = match borsh::to_vec(&headers) {
-        Ok(data) => data,
-        Err(err) => panic_str(&format!("Cannot serialize headers: {:?}", err)),
-    };
-    let method = sys::Buffer::from(method);
-    let url = sys::Buffer::from(url);
-    let headers = sys::Buffer::from(headers.as_slice());
-    let body = sys::Buffer::from(body.as_slice());
-    unsafe { sys::fetch(method, url, headers, body, DATA_REGISTER) }
-    String::from_utf8(read_register(DATA_REGISTER).unwrap_or_else(expected_register)).unwrap()
 }
